Added duplicate_default config
When true, and exists an install filter with these properties:
- Is for a default route
- Has a table
- Has a pref-src
the same route is installed in the selected table and in the export_table.
This allows to have dynamic default routes installed both for output and src-based input response.

To work, it still needs to define manually a route policy from the pref-src to lookup the table, however this can be automated in the future by re-adding rule management to kernel_netlink